- Remove GameAppRunOpt. Engine will run in editor only if compiled in that mode using pre-processor options, as to remove any unwanted code from a game exe. - Updated ConstantUpdate call. We will now call it how it should be called. - Update Timer Params to have the double variables and float, no need to cast in other modules. - Update Angel Script module registration to remove std::wstring reference

+++ CodeSolution/Code/AEngine_GameComponents/Particle System/ParticleEmitter.cpp
@@ -51,7 +51,7 @@ ParticleEmitter::~ParticleEmitter()
 void ParticleEmitter::Update(const TimerParams& timerParams, const XMFLOAT3& newPosition)
 {
 	// Work out how much time has passed since the previous update.
-	float elapsedTime = (float)timerParams.m_ElapsedTime;
+	float elapsedTime = timerParams.m_ElapsedTime;
 
 	if (elapsedTime > 0 && m_ParticleSystem != nullptr)
 	{
